+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.11.0-beta.4
 * fix `helpers.updateSettings`
+* adds optional `connector_name` parameter to `Hull` client and `Hull.Connector` which if set will be added to logs context
+* by default set json format of the logger console output
 
 # 0.11.0-beta.3
 * fix the `requestExtract` handler - allow passing `path` param